requirements.txt
scrapy
scrapy-user-agents
python-dotenv
psycopg2
pandas